Stories done!
* Will now play a pages sound on loading
* Generation of correct/incorrect pages
* Scale is now used when drawing cards on the screen

+++ KiddyApp/DisplayCard.cs
@@ -7,7 +7,7 @@ namespace KiddyApp
     public class DisplayCard
     {
         public Card Card;
-        public double Scale;
+        public float Scale;
         public int X, Y;
 
         public DisplayCard(Card Card)
@@ -16,7 +16,7 @@ namespace KiddyApp
             this.Scale = 1;
         }
 
-        public DisplayCard(Card Card, double Scale)
+        public DisplayCard(Card Card, float Scale)
         {
             this.Card = Card;
             this.Scale = Scale;
