+++ KiddyApp/Kiddy.cs
@@ -7,9 +7,9 @@ namespace KiddyApp
     public class Kiddy
     {
         public string Name = "Celeste";
-        public double Age = 5;
+        public float Age = 5;
         public Difficulty Difficulty = Difficulty.Standard;
-        public double CurrentLevel = 10;
+        public float CurrentLevel = 10;
 
 
         public static Kiddy CurrentPlayer = new Kiddy();
